requirements.txt
discord.py
google-generativeai
python-dotenv
requests